✨ Feat: 팀페이지 - 말줄임 추가, 여백 수정,  할일 목록 수정시 input에 기존값 보이게
